+++ lib/app/app_module.dart
@@ -2,8 +2,7 @@ import 'package:flutter_modular/flutter_modular.dart';
 import 'package:system_maua_front/app/modules/splash/splash_module.dart';
 import 'package:system_maua_front/app/shared/components/bottom_navigation_bar/bottom_navigation_bar_controller.dart';
 
-import 'modules/home-aluno/home_aluno_module.dart';
-import 'modules/notas-aluno/notas_aluno_module.dart';
+import 'modules/landing/landing_module.dart';
 
 class AppModule extends Module {
   @override
@@ -13,8 +12,13 @@ class AppModule extends Module {
 
   @override
   final List<ModularRoute> routes = [
-    ModuleRoute('/home', module: HomeAlunoModule()),
-    ModuleRoute('/splash', module: SplashModule()),
-    ModuleRoute('/notas', module: NotasAlunoModule()),
+    ModuleRoute(
+      '/',
+      module: LandingModule(),
+    ),
+    ModuleRoute(
+      '/splash',
+      module: SplashModule(),
+    ),
   ];
 }
