+++ lib/app/modules/home-aluno/home_aluno_module.dart
@@ -1,5 +1,6 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:system_maua_front/app/modules/home-aluno/repositories/home_aluno_repository.dart';
+import 'package:system_maua_front/app/modules/notas-aluno/notas_aluno_module.dart';
 import 'home_aluno_controller.dart';
 
 import 'home_aluno_page.dart';
@@ -9,11 +10,15 @@ class HomeAlunoModule extends Module {
   @override
   final List<Bind> binds = [
     Bind.lazySingleton((i) => HomeAlunoController(i())),
-    Bind.lazySingleton<IHomeAlunoRepository>((i) => HomeAlunoRepository())
+    Bind.lazySingleton<IHomeAlunoRepository>((i) => HomeAlunoRepository()),
   ];
 
   @override
   final List<ModularRoute> routes = [
-    ChildRoute(Modular.initialRoute, child: (_, args) => HomeAlunoPage()),
+    ChildRoute(Modular.initialRoute,
+        child: (_, args) => HomeAlunoPage(),
+        children: [
+          ModuleRoute('/notas', module: NotasAlunoModule()),
+        ]),
   ];
 }
