+++ lib/app/shared/components/floating_action_button_custom/floating_action_button_custom_widget.dart
@@ -29,8 +29,8 @@ class _FloatingActionButtonCustomWidgetState
   @override
   Widget build(BuildContext context) {
     return SizedBox(
-      width: MediaQuery.of(context).size.width * 0.2,
-      height: MediaQuery.of(context).size.height * 0.2,
+      width: 64,
+      height: 64,
       child: GestureDetector(
         child: FloatingActionButton(
           elevation: 10,
